finishing up upgrades to performance measurement to comply with new plugins api. Changed module parameter change event to imply only rt context can handle.

+++ test/module_tests.cpp
@@ -39,7 +39,7 @@ std::vector<Modules::Variable::Info> generateDefaultComponentVariables()
   tempvarinfo.name = std::string("testuint");
   tempvarinfo.description = std::string("TEST UNSIGNED INTEGER");
   tempvarinfo.vartype = Modules::Variable::UINT_PARAMETER;
-  tempvarinfo.value = 0ull;
+  tempvarinfo.value = 0ul;
   default_variable_list.push_back(tempvarinfo);
 
   tempvarinfo.name = std::string("teststring");
@@ -81,7 +81,7 @@ TEST_F(ModulePluginTests, ComponentParameters)
   ASSERT_DOUBLE_EQ(
       std::get<double>(variable_list[1].value),
       this->plugin->getComponentDoubleParameter(variable_list[1].name));
-  ASSERT_EQ(std::get<unsigned long long>(variable_list[2].value),
+  ASSERT_EQ(std::get<unsigned long>(variable_list[2].value),
             this->plugin->getComponentUIntParameter(variable_list[2].name));
 }
